♻️ Annex tags.page into index.page
